Remove user interaction in activity
Previously user was show all the available high speed camera setting to
chose from. This is bypassed with automatic selection of lowest workable setting.
