+++ Camera2SlowMotion/app/src/main/res/navigation/nav_graph.xml
@@ -22,27 +22,11 @@
 
     <fragment
         android:id="@+id/permissions_fragment"
-        android:name="com.example.android.camera2.slowmo.fragments.PermissionsFragment"
+        android:name="com.example.android.camera2.slowmo.fragments.PermissionsAndCameraSettingFragment"
         android:label="Permissions" >
-
         <action
-            android:id="@+id/action_permissions_to_selector"
-            app:destination="@id/selector_fragment"
-            app:popUpTo="@id/permissions_fragment"
-            app:popUpToInclusive="true" />
-
-    </fragment>
-
-    <fragment
-        android:id="@+id/selector_fragment"
-        android:name="com.example.android.camera2.slowmo.fragments.SelectorFragment"
-        android:label="Selector" >
-
-        <action
-            android:id="@+id/action_selector_to_camera"
-            app:launchSingleTop="true"
-            app:destination="@id/camera_fragment" />
-
+          android:id="@+id/action_permissions_fragment_to_camera_fragment"
+          app:destination="@id/camera_fragment" />
     </fragment>
 
     <fragment
